Fix move probabilities of actions

The function corresponds the provided actions into moves of nature with specified move probabilities. Can be a useful step when checking for existence of equilibria with particular structure.

game_fix_actions(game, ..., actions = list(...), tremble.prob = NULL)

Arguments

...
directly the named arguments from which actions will be constructed

actions
a named list. The names correspond to action names. The default value to fix mixed strategies is a table that specifies conditional move probabilities (see example). If you want to fix pure actions you can also provide arguments as in game_fix_action_preferences.

tremble.prob
If a positive number, we assume that with this probability the player trembles and then chooses a random action with uniform probability. Trembles can be useful to enforce some sequential rationality in continuation play, but note that uniform trembles are not neccessarily the correct form of trembles to find sequential equilibria or trembling hand perfect equilibria.

Details

For fixing pure strategies game_fix_action_preferences is preferable when using the gambit-logit solver that can find sequential equilibria, by using logit trembles.

See also